+++ src/NumSharp.Core/Selection/NDArray.Indexing.Selection.Setter.cs
@@ -458,7 +458,7 @@ namespace NumSharp
 
             //by now all indices are flat, relative indices, might be subshaped, might be non-linear ---------------
             //we flatten to linear absolute points -----------------------------------------------------------------
-            var computedOffsets = new NDArray<int>(Shape.Vector(indicesSize), false);
+            var computedOffsets = new NDArray<long>(Shape.Vector(indicesSize), false);
             var computedAddr = computedOffsets.Address;
 
             //prepare indices getters
@@ -487,7 +487,7 @@ namespace NumSharp
                     for (int ndIdx = 0; ndIdx < ndsCount; ndIdx++) //todo optimize this loop with unmanaged address.
                         index[ndIdx] = indexGetters[ndIdx](i); //replace with memory access or iterators
 
-                    if ((computedAddr[i] = (int)srcShape.GetOffset(index, ndsCount)) > largestOffset)
+                    if ((computedAddr[i] = srcShape.GetOffset(index, ndsCount)) > largestOffset)
                         throw new IndexOutOfRangeException($"Index [{string.Join(", ", new Span<long>(index, ndsCount).ToArray())}] exceeds given NDArray's bounds. NDArray is shaped {srcShape}.");
                 }
             }
@@ -496,7 +496,7 @@ namespace NumSharp
                 var getter = indexGetters[0];
                 for (long i = 0; i < indicesSize; i++)
                 {
-                    if ((computedAddr[i] = (int)srcShape.GetOffset_1D(getter(i))) > largestOffset)
+                    if ((computedAddr[i] = srcShape.GetOffset_1D(getter(i))) > largestOffset)
                         throw new IndexOutOfRangeException($"Index [{getter(i)}] exceeds given NDArray's bounds. NDArray is shaped {srcShape}.");
                 }
             }
@@ -550,7 +550,7 @@ namespace NumSharp
         /// <param name="retShape"></param>
         /// <param name="absolute">Is the given <paramref name="dstOffsets"/> already point to the offset of <paramref name="dst"/>.</param>
         /// <returns></returns>
-        protected static unsafe void SetIndicesND<T>(NDArray<T> dst, NDArray<int> dstOffsets, NDArray[] dstIndices, int ndsCount, long[] retShape, long[] subShape, NDArray<T> values) where T : unmanaged
+        protected static unsafe void SetIndicesND<T>(NDArray<T> dst, NDArray<long> dstOffsets, NDArray[] dstIndices, int ndsCount, long[] retShape, long[] subShape, NDArray<T> values) where T : unmanaged
         {
             Debug.Assert(dstOffsets.size == values.size);
 
@@ -564,7 +564,7 @@ namespace NumSharp
             for (int i = 0; i < subShape.Length; i++)
                 subShapeSize *= subShape[i];
 
-            int* offsetAddr = dstOffsets.Address;
+            long* offsetAddr = dstOffsets.Address;
             long offsetsSize = dstOffsets.size;
             T* valuesAddr = values.Address;
             T* dstAddr = dst.Address;
